Find Anagrams in a huge text. Mapper
This program is a Mapper program. Its function is to go through a text file and create a list of words that are anagrams. The words are stored in the format "Key<tab>Value" in an output file. This output file will be taken as the input to the Reducer program.