Find anagrams in a huge text
This program is a Mapper program. Its function is to go through a text file and create a list of words that are anagrams. The words are stored in the format "Key<tab>Value" in an output file. These Key-Value pairs are stored in the output file after being sorted.