Find users with same DNA sequence -  Mapper
This is a mapper program. It reads an input file which is in the format Username<space>DNAseq.

The program considers the DNA seq as Key and Username as the Value. It sorts the records on the DNA seq and then stores each record in the output file in the format DNAseq<tab>Username