Find users with same DNA seq - Reducer program
This is the Reducer program that displays the list of users with the same DNA sequence grouped together.

The input to this program is a file consisting of tab-separated Key-Value pairs that have already been sorted. Here DNA seq value is the Key and the username is the Value

The program displays each DNA seq and the list of users having that DNA sequence. It also stores the same in an output file.